Add multidestroy concern
