+++ app/controllers/issues_controller.rb
@@ -1,5 +1,6 @@
 class IssuesController < ProjectScopedController
   include ContentFromTemplate
+  include MultipleDestroy
 
   before_action :find_issuelib
   before_action :find_issues, except: [:destroy, :merging]
@@ -87,33 +88,6 @@ class IssuesController < ProjectScopedController
     end
   end
 
-  def multiple_destroy
-    @issues = Issue.where(id: params[:ids])
-
-    # cache these values
-    @count = @issues.count
-    @max_deleted_inline = ::Configuration.max_deleted_inline
-
-    if @count > 0
-      @job_logger = Log.new
-      job_params = {
-        author_email: current_user.email,
-        ids: @issues.map(&:id),
-        klass: 'Issue',
-        uid: @job_logger.uid
-      }
-
-      if @count > @max_deleted_inline
-        @job_logger.write 'Enqueueing multiple delete job to start in the background.'
-        job = MultiDestroyJob.perform_later(job_params)
-        @job_logger.write "Job id is #{job.job_id}."
-      elsif @count > 0
-        @job_logger.write 'Performing multiple delete job inline.'
-        MultiDestroyJob.perform_now(job_params)
-      end
-    end
-  end
-
   def import
     importer = IssueImporter.new(params)
     @results = importer.query()
@@ -169,6 +143,10 @@ class IssuesController < ProjectScopedController
 
   def issue_params
     params.require(:issue).permit(:tag_list, :text)
+  end
+
+  def item_class_name
+    'Issue'
   end
 
   # This method inspect the issues' Tag field and if present tags the issue
